Remove old matching engine
