+++ tomox/batchdb.go
@@ -9,7 +9,6 @@ import (
 	"github.com/ethereum/go-ethereum/ethdb"
 	"github.com/ethereum/go-ethereum/log"
 	lru "github.com/hashicorp/golang-lru"
-	"github.com/pkg/errors"
 )
 
 const (
@@ -24,7 +23,6 @@ type BatchDatabase struct {
 	db           *ethdb.LDBDatabase
 	emptyKey     []byte
 	cacheItems   *lru.Cache // Cache for reading
-	dryRunCaches map[common.Hash]*lru.Cache
 	lock         sync.RWMutex
 	cacheLimit   int
 	Debug        bool
@@ -53,7 +51,6 @@ func NewBatchDatabaseWithEncode(datadir string, cacheLimit int) *BatchDatabase {
 		db:           db,
 		cacheItems:   cacheItems,
 		emptyKey:     EmptyKey(), // pre alloc for comparison
-		dryRunCaches: make(map[common.Hash]*lru.Cache),
 		cacheLimit:   itemCacheLimit,
 	}
 
@@ -75,14 +72,7 @@ func (db *BatchDatabase) HasObject(key []byte, dryrun bool, blockHash common.Has
 	}
 	cacheKey := db.getCacheKey(key)
 
-	if dryrun {
-		db.lock.Lock()
-		dryrunCache, ok := db.dryRunCaches[blockHash]
-		db.lock.Unlock()
-		if ok && dryrunCache.Len() >= 0 && db.dryRunCaches[blockHash].Contains(cacheKey) {
-			return true, nil
-		}
-	} else if db.cacheItems.Contains(cacheKey) {
+	if db.cacheItems.Contains(cacheKey) {
 		// for dry-run mode, do not read cacheItems
 		return true, nil
 	}
@@ -97,16 +87,6 @@ func (db *BatchDatabase) GetObject(key []byte, val interface{}, dryrun bool, blo
 	}
 
 	cacheKey := db.getCacheKey(key)
-	if dryrun {
-		db.lock.Lock()
-		dryrunCache, ok := db.dryRunCaches[blockHash]
-		db.lock.Unlock()
-		if ok && dryrunCache.Len() >= 0 {
-			if value, ok := db.dryRunCaches[blockHash].Get(cacheKey); ok {
-				return value, nil
-			}
-		}
-	}
 
 	// for dry-run mode, do not read cacheItems
 	if cached, ok := db.cacheItems.Get(cacheKey); ok && !dryrun {
@@ -140,19 +120,6 @@ func (db *BatchDatabase) GetObject(key []byte, val interface{}, dryrun bool, blo
 
 func (db *BatchDatabase) PutObject(key []byte, val interface{}, dryrun bool, blockHash common.Hash) error {
 	cacheKey := db.getCacheKey(key)
-	if dryrun {
-		db.lock.Lock()
-		dryrunCache, ok := db.dryRunCaches[blockHash]
-		db.lock.Unlock()
-		if !ok {
-			log.Debug("BatchDB - PutObject: DryrunCache of this block is not initialized. Initialize now!", "blockHash", blockHash)
-			db.InitDryRunMode(blockHash)
-			dryrunCache, _ = db.dryRunCaches[blockHash]
-		}
-		dryrunCache.Add(cacheKey, val)
-		return nil
-	}
-
 	db.cacheItems.Add(cacheKey, val)
 	value, err := EncodeBytesItem(val)
 	if err != nil {
@@ -166,96 +133,9 @@ func (db *BatchDatabase) DeleteObject(key []byte, dryrun bool, blockHash common.
 	// for better performance, we can mark a Deleted flag, to do batch delete
 	cacheKey := db.getCacheKey(key)
 
-	//mark it to nil in dryrun cache
-	if dryrun {
-		db.lock.Lock()
-		dryrunCache, ok := db.dryRunCaches[blockHash]
-		db.lock.Unlock()
-		if !ok {
-			log.Debug("BatchDB - DeleteObject: DryrunCache of this block is not initialized. Initialize now!", "blockHash", blockHash)
-			db.InitDryRunMode(blockHash)
-			dryrunCache, _ = db.dryRunCaches[blockHash]
-		}
-		dryrunCache.Add(cacheKey, nil)
-		return nil
-	}
 
 	db.cacheItems.Remove(cacheKey)
 	return db.db.Delete(key)
-}
-
-func (db *BatchDatabase) InitDryRunMode(blockHash common.Hash) {
-	log.Debug("Start dry-run mode, clear old data", "blockhash", blockHash)
-	db.lock.Lock()
-	dryrunCache, ok := db.dryRunCaches[blockHash]
-	db.lock.Unlock()
-
-	// if the dryrunCache of this blockHash already existed, purge it
-	// otherwise, initialize new cache for it
-	// Finally, assign the cache to db.dryRunCaches
-	if ok && dryrunCache != nil {
-		dryrunCache.Purge()
-	} else {
-		dryrunCache, _ = lru.New(db.cacheLimit)
-	}
-	db.lock.Lock()
-	db.dryRunCaches[blockHash] = dryrunCache
-	db.lock.Unlock()
-}
-
-func (db *BatchDatabase) SaveDryRunResult(blockHash common.Hash) error {
-	log.Debug("Start saving dry-run result to DB ", "blockhash", blockHash)
-	db.lock.Lock()
-	dryrunCache, ok := db.dryRunCaches[blockHash]
-	db.lock.Unlock()
-	if !ok || dryrunCache.Len() == 0 {
-		log.Debug("Nothing to SaveDryRunResult. DryrunCache is empty.", "blockhash", blockHash)
-		return nil
-	}
-	batch := db.db.NewBatch()
-	for _, cacheKey := range dryrunCache.Keys() {
-		key, err := hex.DecodeString(cacheKey.(string))
-		if err != nil {
-			log.Error("Can't save dry-run result (hex.DecodeString)", "err", err)
-			return err
-		}
-		val, ok := dryrunCache.Get(cacheKey)
-		if !ok {
-			err := errors.New("can't get item from dryrun cache")
-			log.Error("Can't save dry-run result (db.dryRunCache.GetObject)", "err", err)
-			return err
-		}
-		if val == nil {
-			if err := db.db.Delete(key); err != nil {
-				log.Error("Can't save dry-run result (db.db.DeleteObject)", "err", err)
-				return err
-			}
-			continue
-		}
-
-		value, err := EncodeBytesItem(val)
-		if err != nil {
-			log.Error("Can't save dry-run result (EncodeBytesItem)", "err", err)
-			return err
-		}
-
-		if err := batch.Put(key, value); err != nil {
-			log.Error("Can't save dry-run result (batch.PutObject)", "err", err)
-			return err
-		}
-	}
-	log.Debug("Successfully saved dry-run result to DB ", "blockhash", blockHash)
-	dryrunCache.Purge()
-	db.lock.Lock()
-	delete(db.dryRunCaches, blockHash)
-	db.lock.Unlock()
-	// purge reading cache to refresh data from db
-	db.cacheItems.Purge()
-	return batch.Write()
-}
-
-func (db *BatchDatabase) CancelOrder(hash common.Hash) error {
-	return nil
 }
 
 func (db *BatchDatabase) Put(key []byte, val []byte) error {
