--- tomox/iterator.go
@@ -1,161 +1,0 @@
-// Copyright (c) 2019, Agiletech Viet Nam. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package tomox
-
-import "github.com/ethereum/go-ethereum/common"
-
-// Iterator holding the iterator's state
-type Iterator struct {
-	tree     *Tree
-	node     *Node
-	position position
-}
-
-type position byte
-
-const (
-	begin, between, end position = 0, 1, 2
-)
-
-// Iterator returns a stateful iterator whose elements are key/value pairs.
-func (tree *Tree) Iterator() Iterator {
-	return Iterator{tree: tree, node: nil, position: begin}
-}
-
-// Next moves the iterator to the next element and returns true if there was a next element in the container.
-// If Next() returns true, then next element's key and value can be retrieved by Key() and Value().
-// If Next() was called for the first time, then it will point the iterator to the first element if it exists.
-// Modifies the state of the iterator.
-func (iterator *Iterator) Next(dryrun bool, blockHash common.Hash) bool {
-	if iterator.position == end {
-		goto end
-	}
-	if iterator.position == begin {
-		left := iterator.tree.Left(dryrun, blockHash)
-		if left == nil {
-			goto end
-		}
-		iterator.node = left
-		goto between
-	}
-	if len(iterator.node.Item.Keys.Right) > 0 {
-		iterator.node = iterator.node.Right(iterator.tree, dryrun, blockHash)
-		if iterator.node == nil {
-			goto end
-		}
-		for len(iterator.node.Item.Keys.Left) > 0 {
-			iterator.node = iterator.node.Left(iterator.tree, dryrun, blockHash)
-			if iterator.node == nil {
-				goto end
-			}
-		}
-		goto between
-	}
-	if len(iterator.node.Item.Keys.Parent) > 0 {
-		node := iterator.node
-		for len(iterator.node.Item.Keys.Parent) > 0 {
-			if iterator.node == nil {
-				goto end
-			}
-			iterator.node = iterator.node.Parent(iterator.tree, dryrun, blockHash)
-			if iterator.tree.Comparator(node.Key, iterator.node.Key) <= 0 {
-				goto between
-			}
-		}
-	}
-
-end:
-	iterator.node = nil
-	iterator.position = end
-	return false
-
-between:
-	iterator.position = between
-	return true
-}
-
-// Prev moves the iterator to the previous element and returns true if there was a previous element in the container.
-// If Prev() returns true, then previous element's key and value can be retrieved by Key() and Value().
-// Modifies the state of the iterator.
-func (iterator *Iterator) Prev(dryrun bool, blockHash common.Hash) bool {
-	if iterator.position == begin {
-		goto begin
-	}
-	if iterator.position == end {
-		right := iterator.tree.Right(dryrun, blockHash)
-		if right == nil {
-			goto begin
-		}
-		iterator.node = right
-		goto between
-	}
-	if iterator.node.Item.Keys.Left != nil {
-		iterator.node = iterator.node.Left(iterator.tree, dryrun, blockHash)
-		for iterator.node.Item.Keys.Right != nil {
-			iterator.node = iterator.node.Right(iterator.tree, dryrun, blockHash)
-		}
-		goto between
-	}
-	if iterator.node.Item.Keys.Parent != nil {
-		node := iterator.node
-		for iterator.node.Item.Keys.Parent != nil {
-			iterator.node = iterator.node.Parent(iterator.tree, dryrun, blockHash)
-			if iterator.tree.Comparator(node.Key, iterator.node.Key) >= 0 {
-				goto between
-			}
-		}
-	}
-
-begin:
-	iterator.node = nil
-	iterator.position = begin
-	return false
-
-between:
-	iterator.position = between
-	return true
-}
-
-// Value returns the current element's value.
-// Does not modify the state of the iterator.
-func (iterator *Iterator) Value() []byte {
-	return iterator.node.Item.Value
-}
-
-// Key returns the current element's key.
-// Does not modify the state of the iterator.
-func (iterator *Iterator) Key() []byte {
-	return iterator.node.Key
-}
-
-// Begin resets the iterator to its initial state (one-before-first)
-// Call Next() to fetch the first element if any.
-func (iterator *Iterator) Begin() {
-	iterator.node = nil
-	iterator.position = begin
-}
-
-// End moves the iterator past the last element (one-past-the-end).
-// Call Prev() to fetch the last element if any.
-func (iterator *Iterator) End() {
-	iterator.node = nil
-	iterator.position = end
-}
-
-// First moves the iterator to the first element and returns true if there was a first element in the container.
-// If First() returns true, then first element's key and value can be retrieved by Key() and Value().
-// Modifies the state of the iterator
-func (iterator *Iterator) First(dryrun bool, blockHash common.Hash) bool {
-	iterator.Begin()
-	return iterator.Next(dryrun, blockHash)
-}
-
-// Last moves the iterator to the last element and returns true if there was a last element in the container.
-// If Last() returns true, then last element's key and value can be retrieved by Key() and Value().
-// Modifies the state of the iterator.
-func (iterator *Iterator) Last(dryrun bool, blockHash common.Hash) bool {
-	iterator.End()
-	return iterator.Prev(dryrun, blockHash)
-}
