+++ tomox/order.go
@@ -1,14 +1,10 @@
 package tomox
 
 import (
-	"bytes"
 	"fmt"
 	"github.com/ethereum/go-ethereum/tomox/tomox_state"
 	"math/big"
 	"time"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/log"
 )
 
 const (
@@ -59,51 +55,4 @@ func (order *Order) String() string {
 
 	return fmt.Sprintf("orderID : %s, price: %s, quantity :%s, relayerID: %s",
 		new(big.Int).SetBytes(order.Key), order.Item.Price, order.Item.Quantity, order.Item.ExchangeAddress.Hex())
-}
-
-func (order *Order) GetNextOrder(orderList *OrderList, dryrun bool, blockHash common.Hash) *Order {
-	nextOrder := orderList.GetOrder(order.Item.NextOrder, dryrun, blockHash)
-
-	return nextOrder
-}
-
-func (order *Order) GetPrevOrder(orderList *OrderList, dryrun bool, blockHash common.Hash) *Order {
-	prevOrder := orderList.GetOrder(order.Item.PrevOrder, dryrun, blockHash)
-
-	return prevOrder
-}
-
-// NewOrder : create new order with quote ( can be ethereum address )
-func NewOrder(orderItem *tomox_state.OrderItem, orderListKey []byte) *Order {
-	key := GetKeyFromBig(new(big.Int).SetUint64(orderItem.OrderID))
-	orderItem.NextOrder = EmptyKey()
-	orderItem.PrevOrder = EmptyKey()
-	orderItem.OrderList = orderListKey
-	// key should be Hash for compatible with smart contract
-	order := &Order{
-		Key:  key,
-		Item: orderItem,
-	}
-
-	return order
-}
-
-// UpdateQuantity : update quantity of the order
-func (order *Order) UpdateQuantity(orderList *OrderList, newQuantity *big.Int, dryrun bool, blockHash common.Hash) error {
-	if newQuantity.Cmp(order.Item.Quantity) > 0 && !bytes.Equal(orderList.Item.TailOrder, order.Key) {
-		if err := orderList.MoveToTail(order, dryrun, blockHash); err != nil {
-			return err
-		}
-	}
-	// update volume and modified timestamp
-	orderList.Item.Volume = Sub(orderList.Item.Volume, Sub(order.Item.Quantity, newQuantity))
-	order.Item.Quantity = CloneBigInt(newQuantity)
-	log.Debug("QUANTITY", order.Item.Quantity.String())
-	if err := orderList.SaveOrder(order, dryrun, blockHash); err != nil {
-		return err
-	}
-	if err := orderList.Save(dryrun, blockHash); err != nil {
-		return err
-	}
-	return nil
 }
